fix(embedded): 真机取证修正蓝灯断言口径，并补 IDF 5.5.5 版 ADF FreeRTOS 补丁
真机（COM12，ESP32-S3-WROOM-1-N16R8）首轮烧录后暴露两处问题，均已修并复测：

1. 板级自检蓝灯误报
   ADF get_blue_led_gpio() 返回 int8_t，TCA9554 P7 = BIT(7) = 0x80 经 int8_t 回传为 -128；
   原断言期望 128 ⇒ 真机 FAIL（板子无问题，是取值类型口径）。
   改为按同口径折算 (int)(int8_t)(1 << 7)，并在 board_expect.h 注明。

2. 6 键任务创建失败（ADF 侧必需补丁）
   ADF audio_thread 需 xTaskCreateRestrictedPinnedToCore()（静态栈 + 指定核），
   IDF v5.5.5 未提供 ⇒ 串口 "Not found right xTaskCreateRestrictedPinnedToCore"
   → "ADC_BTN: Create button_task failed!"，按键全不可用。
   上游 idf_patches/idf_v5.5_freertos.patch 面向 v5.5.0，在本栈（v5.5.5）git apply --check 失败；
   故新增适配版 idf_patches/idf_v5.5.5_freertos.patch（实现 + 声明 + linker_common.lf 片段，
   声明置于 configSUPPORT_STATIC_ALLOCATION 段内），已在 WSL IDF v5.5.5 实打并复测。

真机复测结果（重启后首启日志）：
- 板级自检 19 项 / 失败 0（STRICT 不再中止）
- ADC_BTN 校准成功 + "按键服务已启动（6 键 → event/up）"
- SD 卡 1-line 挂载 PASS（CID name SD!）、SPIFFS 兜底 934 KB
- AEC 采集就绪（16000 Hz / 32 bit / RMNM → AFE → WAV）、板上回放就绪
- Octal PSRAM 8MB @80MHz 确认（与 CONFIG_SPIRAM_MODE_OCT=y 假设一致）；flash 实测 16 MB
- 凭据文件配网：SD 根目录暂无 oneye-wifi.txt ⇒ 按设计降级（不中止）并打印投放指引

+++ examples/oneye/korvo2_oneye/main/korvo2_oneye_main.c
@@ -137,7 +137,9 @@ static void board_selftest(void)
     chk_int("sdcard_pwr_ctrl_gpio", -1, get_sdcard_power_ctrl_gpio());
     chk_int("sdcard_open_file_num_max", 5, get_sdcard_open_file_num_max());
     chk_int("green_led_gpio (无)", -1, get_green_led_gpio());
-    chk_int("blue_led_gpio (TCA9554 P7)", (1 << 7), get_blue_led_gpio());
+    /* ADF 的 getter 返回 **int8_t**，而 TCA9554 P7 位 = BIT(7) = 0x80 ⇒ 经 int8_t 回传为 -128。
+     * 真机实测即 -128（不是板子的问题，是取值类型口径）：期望值按同口径折算再比对。 */
+    chk_int("blue_led_gpio (TCA9554 P7)", (int)(int8_t)(1 << 7), get_blue_led_gpio());
     chk_int("es8311_mclk_src (ESP MCLK)", 0, get_es8311_mclk_src());
 
     /* ADC 输入通道格式：rst「AEC 电路」——4 通道，R = 硬件 AEC 回采；board_def.h:126 */
